feat: split DataTable in several modules + tests
