+++ src/types/Column.ts
@@ -1,0 +1,10 @@
+/**
+ * Column interface
+ * @property {number | string} id - the id of the cell
+ * @property {string} label - the label of the cell
+ */
+export interface Column {
+  id: number | string;
+  label: string;
+  render?: ((value: any) => JSX.Element) | null;
+}
